requirements.txt
pip~=24.3.1
attrs~=25.1.0
idna~=3.10
multidict~=6.1.0
propcache~=0.2.1
yarl~=1.18.3
aiohttp~=3.11.12
aiosignal~=1.3.2
frozenlist~=1.5.0
aiohappyeyeballs~=2.4.6
disnake~=2.10.1
python-dotenv~=1.0.1